* RGB16f is half not float... * RGB16f id is Rgb16f nowdays
svn path=/trunk/koffice/; revision=855390

+++ krita/colorspaces/rgb_float_hdr.ctl/kis_rgb_f16_hdr_colorspace.h
@@ -25,7 +25,7 @@
 
 #include "kis_rgb_float_hdr_colorspace.h"
 
-typedef KoRgbTraits<float> RgbF16Traits;
+typedef KoRgbTraits<half> RgbF16Traits;
 
 class KisRgbF16HDRColorSpace : public KisRgbFloatHDRColorSpace<RgbF16Traits>
 {
